+++ src/stonks_overwatch/services/brokers/bitvavo/repositories/assets_repository.py
@@ -1,11 +1,11 @@
-from django.db import connection
-
-from stonks_overwatch.utils.database.db_utils import dictfetchone
+from stonks_overwatch.services.brokers.bitvavo.repositories.models import BitvavoAssets
+from stonks_overwatch.utils.database.db_utils import dictfetchone, get_connection_for_model
 
 
 class AssetsRepository:
     @staticmethod
     def get_asset(symbol: str) -> dict | None:
+        connection = get_connection_for_model(BitvavoAssets)
         with connection.cursor() as cursor:
             cursor.execute(
                 """