+++ src/stonks_overwatch/services/brokers/bitvavo/repositories/transactions_repository.py
@@ -1,11 +1,11 @@
-from django.db import connection
-
-from stonks_overwatch.utils.database.db_utils import dictfetchall
+from stonks_overwatch.services.brokers.bitvavo.repositories.models import BitvavoTransactions
+from stonks_overwatch.utils.database.db_utils import dictfetchall, get_connection_for_model
 
 
 class TransactionsRepository:
     @staticmethod
     def get_transactions_raw() -> list[dict]:
+        connection = get_connection_for_model(BitvavoTransactions)
         with connection.cursor() as cursor:
             cursor.execute(
                 """
@@ -17,6 +17,7 @@ class TransactionsRepository:
 
     @staticmethod
     def get_deposits_history_raw() -> list[dict]:
+        connection = get_connection_for_model(BitvavoTransactions)
         with connection.cursor() as cursor:
             cursor.execute(
                 """